Fix unaligned access in buffer copies.
There's a TODO in the source about fixing alignment of the address being
copied.  This unaligned access is a problem on armhf with some kernel
configurations, and is generally slower even on other architectures due
to the fixups.  Just use memcpy instead which is bound to do the right
thing.

This fixes the Debian bug #832069.